requirements.dev.txt
# dev dependencies
ipython
ipdb
pytest-watch

# Install project as editable
-e .